# Minor optimization in hashable_partition