This is just a very small change to `hashable_partition` that avoids calling `filter_fn` twice for every leaf.

On my machine this saves 200μs on the MLP example (~260 leaves).

All tests are passing and there should be no behavior change.

Old: 

```Python
%timeit -r7 -n500 fwd_eqx(dyn_flat, x).block_until_ready()
2.36 ms ± 58.1 μs per loop (mean ± std. dev. of 7 runs, 500 loops each)
```

New:

```Python
%timeit -r7 -n500 fwd_eqx(dyn_flat, x).block_until_ready()
2.16 ms ± 47 μs per loop (mean ± std. dev. of 7 runs, 500 loops each)
```